requirements.txt
fastapi
uvicorn[standard]
Pillow
python-multipart
jinja2
docx2pdf
PyPDF2